+++ jsolex/src/main/resources/whats-new.md
@@ -17,6 +17,7 @@
 - Fixed autostretch to be less aggressive on prominences, providing a better transition between the disk and the prominences
 - Fixed angle P correction being wrongly applied to spectrum images in flare/ellerman bombs detection
 - Process 2 files in parallel in batch mode
+- Fixed position of ellipse when image is rescaled
 
 ### 3.2.0